Implement all missing Rules and adjust utilities and entities to match reassignment and different evaluation, as we no longer use Optional

+++ src/main/kotlin/com/example/kotlineventprocessor/producer/SapRequest.kt
@@ -1,11 +1,12 @@
 package com.example.kotlineventprocessor.producer
 
+import java.math.BigDecimal
 import java.time.LocalDateTime;
 
 data class SapRequest(
     val businessPartnerNumber : String,
     val contractId : String,
-    var amount: Double,
+    var amount: BigDecimal,
     val mProcess: String?,
     val sProcess: String?,
     val bankRoutingNumber : String,
